Add a simple follow button example
Added a very very simple AJAX follow button example that uses the
interface to perform the call.  Uses no DB interaction, so this is NOT
an optimal way of storing or grabbing the data.  This was put in as an
example to show that the interface, when used properly, can use AJAX
calls to perform actions almost identical to how Twitch handles the
function on their site.