Fix to Iteration code
Added a check to the iteration code that will allow the interface to
detect if a return came back improper (A key not existing in the array)
and retry the call.  Respects the limit of retries allowed.  Also added
default values to $limit and $offset in the chat_getEmoticons() and
chat_getEmoticonsGlobal() calls, allowing you to omit those params since
they are optional (and as of now, actually useless.)