[IN TESTING] Commit untested total returns
Changes all functions that use the iteration function to allow the
return of a _total row in the array.  This, in limitless calls, is the
actual count of all returned rows (For cases where data changes during
the call, I didn't want to try and figure out where the data would be
and recall it.  Would be too easy to hit script execution time.).  I
limited calls, pulls directly from the _total object if it exists.