SSL certificate changes
Updates #5.  Implement the ability for the interface to properly use an
SSL certificate in calls.  If OpenSSL is installed and available, will
validate your certificate before using it and enabling SSL and will
attempt to generate a CRT and a self-signed certificate for use in SSL.
(The checks needed before generation are still being written.  As of
now, it will generate a blank CRT and will write a failing certificate).
Many new config options added.  Documentation will be updated as soon as
the system is 100% and completed.