Close #5
OpenSSL removed completely from the interface.  Between its lack of
proper documentation and the simple fact that it does NOT currently have
any portability or default usability (Windows users are SOL and Linux
still requires a lot of configuration) it is not a solution to
generating a certificate.  if you have a certificate, use the new config
option to define your cert to use.  Validation is on you.